task_page_create

+++ server/routes/api/api.items.routes.js
@@ -17,11 +17,11 @@ router.get('/', async (req, res) => {
 
   router.post('/', async (req, res) => {
     try {
-      const {description, price, user_id } = req.body;
+      const {description, price } = req.body;
       //тут создание галлереии и картинки
-      const item = await Item.create({description,price,user_id,});
+      const item = await Item.create({description,price,user_id:res.locals.user.id,});
       const createGallery = await ItemGallery.create({item_id:item.id})
-      const newItem = await Item.findOne({where: { user_id,id:item.id }, include: {
+      const newItem = await Item.findOne({where: { user_id:res.locals.user.id,id:item.id }, include: {
         model: ItemGallery,
         include: ItemImage
       }});
@@ -35,7 +35,7 @@ router.get('/', async (req, res) => {
   router.delete('/:itemId', async (req, res) => {
     try {
       const { itemId } = req.params;
-      const result = await Item.destroy({ where: { id: itemId } });
+      const result = await Item.destroy({ where: { id: itemId, user_id:res.locals.user.id } });
       if (result > 0) {
         res.json({ message: 'success', itemId });
         return;
